:spark:: 이름 변경 mpcManager -> ExchangeViewModel
